% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11-package.R
\docType{package}
\name{cpp11-package}
\alias{cpp11}
\alias{cpp11-package}
\title{cpp11: A C++11 Interface for R's C Interface}
\description{
Provides a header only, C++11 interface to R's C interface.
  Compared to other approaches 'cpp11' strives to be safe against long jumps
  from the C API as well as C++ exceptions, conform to Rs normal semantics and
  supports interaction with 'ALTREP' vectors.
}
\author{
\strong{Maintainer}: Jim Hester \email{james.f.hester@gmail.com} (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
