% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{cpp_generate_bindings}
\alias{cpp_generate_bindings}
\title{Generate C function exports and R wrappers for decorated functions}
\usage{
cpp_generate_bindings(path = ".")
}
\arguments{
\item{path}{The path to the package root directory}
}
\value{
The paths to the generated R and C++ source files (in that order).
}
\description{
Functions decorated with \verb{[[cpp11::export]]} in files ending in \code{.cc},
\code{.cpp}, \code{.h} or \code{.hpp} will be wrapped in generated code which allows them to be
called from R.
}
