% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unvendor.R
\name{cpp_unvendor}
\alias{cpp_unvendor}
\title{Unvendor the cpp11 dependency}
\usage{
cpp_unvendor(path = ".")
}
\arguments{
\item{path}{The path to the package root directory}
}
\value{
The file path to the unvendored code (invisibly).
}
\description{
This function removes the vendored cpp11 headers from your package and
restores the \code{LinkingTo: cpp11} field in the DESCRIPTION file if it was removed.
}
\examples{
# create a new directory
dir <- tempfile()
dir.create(dir)

# vendor the cpp11 headers into the directory
cpp_vendor(dir)

# unvendor the cpp11 headers from the directory
cpp_unvendor(dir)

list.files(file.path(dir, "inst", "include", "cpp11"))

# cleanup
unlink(dir, recursive = TRUE)
}
