% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegen.R
\name{generate_exports}
\alias{generate_exports}
\title{Generate C function exports and R wrappers for decorated functions}
\usage{
generate_exports(path = ".")
}
\arguments{
\item{path}{The path to the package root directory}
}
\description{
Generate C function exports and R wrappers for decorated functions
}
