% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp11-package.R
\docType{package}
\name{cpp11-package}
\alias{cpp11}
\alias{cpp11-package}
\title{cpp11: A C++11 Interface for R's C Interface}
\description{
Provides a header only, C++11 interface to R's C interface. Compared to other approaches 'cpp11' strives to be safe against long jumps from the C API as well as C++ exceptions, conform to normal R function semantics and supports interaction with 'ALTREP' vectors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cpp11.r-lib.org}
  \item \url{https://github.com/r-lib/cpp11}
  \item Report bugs at \url{https://github.com/r-lib/cpp11/issues}
}

}
\author{
\strong{Maintainer}: Romain François \email{romain@posit.co}

Authors:
\itemize{
  \item Jim Hester (\href{https://orcid.org/0000-0002-2739-7082}{ORCID})
}

Other contributors:
\itemize{
  \item Benjamin Kietzman [contributor]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
